+++ product-service/src/main/java/com/palja/product_service/application/service/impl/ProductServiceImpl.java
@@ -129,7 +129,7 @@ public class ProductServiceImpl implements ProductService {
     @Override
     @Transactional
     public UpdateStockRes updateStock(UUID productId,
-                                      Integer stock) {
+                                      Long stock) {
 
         Product product = repository.findProduct(productId);
         CompanyUserInfoRes myInfo = userClient.getMyInfo();
@@ -148,7 +148,7 @@ public class ProductServiceImpl implements ProductService {
     @Override
     @Transactional
     public SaleProductRes saleProductV1(UUID productId,
-                                        Integer quantity) {
+                                        Long quantity) {
 
         Product product = repository.findByIdFetchStockWithLock(productId, quantity);
         product.decreaseStock(quantity);
@@ -158,7 +158,7 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public SaleProductRes saleProduct(UUID productId,
-                                      Integer quantity) {
+                                      Long quantity) {
 
         ProductStock stock = repository.findProductStock(productId);
 
@@ -172,7 +172,7 @@ public class ProductServiceImpl implements ProductService {
     @Override
     @Transactional
     public RestoreStockRes stockRestoreV1(UUID productId,
-                                          Integer quantity) {
+                                          Long quantity) {
 
         Product product = repository.findByIdFetchStockWithLock(productId, quantity);
         product.increaseStock(quantity);
@@ -183,7 +183,7 @@ public class ProductServiceImpl implements ProductService {
     @Override
     @Transactional
     public RestoreStockRes stockRestore(UUID productId,
-                                        Integer quantity) {
+                                        Long quantity) {
 
         ProductStock restoredStock = repository.findProduct(productId).increaseStock(quantity);
 
@@ -197,7 +197,7 @@ public class ProductServiceImpl implements ProductService {
     @Override
     @Transactional
     public DecreaseStockForTimeDealRes decreaseStockForTimeDeal(UUID productId,
-                                                                Integer quantity) {
+                                                                Long quantity) {
 
         Product product = repository.findProduct(productId);
         product.decreaseStock(quantity);
@@ -212,7 +212,7 @@ public class ProductServiceImpl implements ProductService {
     @Override
     @Transactional
     public IncreaseStockForTimeDealRes increaseStockForTimeDeal(UUID productId,
-                                                                Integer quantity) {
+                                                                Long quantity) {
 
         Product product = repository.findProduct(productId);
         product.increaseStock(quantity);
